srv/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
python-jose==3.5.0
passlib[bcrypt]==1.7.4
pydantic==2.11.7